+++ lib/generators/money_attribute/initializer_generator.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
-module Mint
+module MoneyAttribute
   module Generators
     class InitializerGenerator < ::Rails::Generators::Base
       source_root File.expand_path('../templates', __dir__)
 
-      desc 'Creates Minting initializer.'
+      desc 'Creates MoneyAttribute initializer.'
 
       def copy_initializer
-        copy_file 'minting.rb', 'config/initializers/minting.rb'
+        copy_file 'money_attribute.rb', 'config/initializers/money_attribute.rb'
       end
     end
   end