+++ lib/money_attribute.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'minting'
+require 'money_attribute/core_ext'
+require 'money_attribute/configuration'
+require 'money_attribute/macro'
+require 'money_attribute/parser'
+require 'money_attribute/type'
+require 'money_attribute/railtie'
+require 'money_attribute/version'