+++ lib/money_attribute/core_ext.rb
@@ -3,15 +3,15 @@
 # :nodoc
 class Numeric
   def to_money(currency)
-    Mint.money(self, currency)
+    ::Mint.money(self, currency)
   end
 
   def dollars
-    Mint.money(self, 'USD')
+    ::Mint.money(self, 'USD')
   end
 
   def euros
-    Mint.money(self, 'EUR')
+    ::Mint.money(self, 'EUR')
   end
 
   alias dollar dollars
@@ -22,7 +22,7 @@ end
 # :nodoc
 class String
   def to_money(currency)
-    Mint.money(to_r, currency)
+    ::Mint.money(to_r, currency)
   end
 
   alias mint to_money